+++ examples/tinkertoy/TinkertoyWidget.cpp
@@ -39,6 +39,7 @@
 #include "TinkertoyWidget.hpp"
 
 #include "TinkertoyWorldNode.hpp"
+#include "dart/gui/osg/ImGuiHandler.hpp"
 
 //==============================================================================
 TinkertoyWidget::TinkertoyWidget(
@@ -55,8 +56,10 @@ TinkertoyWidget::TinkertoyWidget(
 //==============================================================================
 void TinkertoyWidget::render()
 {
-  ImGui::SetNextWindowPos(ImVec2(10, 20));
-  ImGui::SetNextWindowSize(ImVec2(360, 640));
+  const auto guiScale
+      = static_cast<float>(mViewer->getImGuiHandler()->getGuiScale());
+  ImGui::SetNextWindowPos(ImVec2(10 * guiScale, 20 * guiScale));
+  ImGui::SetNextWindowSize(ImVec2(360 * guiScale, 640 * guiScale));
   ImGui::SetNextWindowBgAlpha(0.5f);
   if (!ImGui::Begin(
           "Tinkertoy Control",